Fortran and model saving cleanup (#43)
Cleans up a few aspects (especially in the Fortran) code to the Fortran
testing applications. No new functionality is added, most of this is
for additional comments and code maintanability

Additionally, the model_saver utility was updated to accept an arbitrary
pretrained Torch model. This will in the future allow us to benchmark
different size models and different architectures.

[ committed by @ashao ]
[ reviewed by @amandarichardsonn ]